+++ pollux/lib/project/Dependencies.scala
@@ -39,9 +39,12 @@ object Dependencies {
 
   private lazy val flyway = "org.flywaydb" % "flyway-core" % Versions.flyway
 
-  private lazy val quillJdbcZio = "io.getquill" %% "quill-jdbc-zio" % Versions.quill
-  private lazy val quillDoobie =
-    "io.getquill" %% "quill-doobie" % Versions.quill exclude ("org.scala-lang.modules", "scala-java8-compat_3")
+  private lazy val quillJdbcZio = "io.getquill" %% "quill-jdbc-zio" %
+    Versions.quill exclude ("org.scala-lang.modules", "scala-java8-compat_3")
+
+  private lazy val quillDoobie = "io.getquill" %% "quill-doobie" %
+    Versions.quill exclude ("org.scala-lang.modules", "scala-java8-compat_3")
+
   private lazy val testcontainers =
     "com.dimafeng" %% "testcontainers-scala-postgresql" % Versions.testContainersScalaPostgresql % Test